# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "Just example project",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Add new account",
                "operationId": "692b5e9720a203958c7433bb5487c072",
                "requestBody": {
                    "$ref": "#/components/requestBodies/SignUpRequest"
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/auth/signin": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Retrieve access token",
                "operationId": "445066a2ac1ab443d1b81af7fef9b961",
                "requestBody": {
                    "$ref": "#/components/requestBodies/SignInRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TokenResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/auth/getaccount": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get account data",
                "operationId": "e24abf371e10f47403cae9f3c7cb23d3",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product index",
                "operationId": "19bfe44c5a4c96ea4c18170ceea0d0a5",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ProductPropertyResource": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Some name"
                    },
                    "value": {
                        "example": "Some value"
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "count": {
                        "type": "integer"
                    },
                    "cost": {
                        "type": "number",
                        "format": "float"
                    },
                    "properties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductPropertyResource"
                        }
                    }
                },
                "type": "object"
            },
            "TokenResource": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "example": "1|VBLj1uXCcemEcHJnaF9nESwLtw4QFCR5vb9F9PJy104027c9"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Smith"
                    },
                    "email": {
                        "type": "string",
                        "example": "test@email.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+79876543232"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "ProductIndexRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "eum sunt"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "SignInRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "login": {
                                    "type": "string",
                                    "example": "test@example.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "password1"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "SignUpRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "John Smith"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "test@example.com"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "+79876543232"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "password1"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "Bearer token": {
                "type": "http",
                "description": "Get one from /api/v1/auth/signin",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "SignUp and SignIn"
        },
        {
            "name": "Products",
            "description": "Product catalog"
        }
    ]
}